refactor: 일급 컬렉션 상태를 유지할 수 있도록 RequestHeaders entity 생성
변경전: HttpRequest entity에서 Collection형태(Hash Map)로 RequestHeaders를 관리
변경후: Collection을 wrapping하고 있는 RequestHeaders entity를 생성해서 일급 컬렉션을 적용
