+++ src/main/java/request/http/HttpRequest.java
@@ -1,4 +1,4 @@
-package request;
+package request.http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
